requirements.txt
python-dotenv~=1.0.1
pytz~=2024.2
SQLAlchemy~=2.0.36
python-telegram-bot~=21.7
python-telegram-bot-calendar~=1.0.5